+++ packages/editor-common/web/src/Components/InlineToolbarButton.jsx
@@ -134,7 +134,6 @@ class InlineToolbarButton extends Component {
       </div>
     );
     /* eslint-enable jsx-a11y/no-static-element-interactions */
-
     return <ToolbarButton theme={theme} tooltipText={tooltipText} button={codeBlockButton} />;
   }
 }